+++ globook_server/src/main/java/org/gdsc/globook/application/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package org.gdsc.globook.application.repository;
 
 import java.util.Optional;
-import org.gdsc.globook.domain.User;
+import org.gdsc.globook.domain.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
