+++ globook_server/src/main/java/org/gdsc/globook/application/service/AuthService.java
@@ -13,10 +13,10 @@ import org.gdsc.globook.core.util.JwtUtil;
 import org.gdsc.globook.core.util.OAuth2Util;
 import org.gdsc.globook.core.util.RandomNicknameUtil;
 import org.gdsc.globook.core.util.RandomPasswordUtil;
-import org.gdsc.globook.domain.ELanguage;
-import org.gdsc.globook.domain.ELoginProvider;
-import org.gdsc.globook.domain.EUserRole;
-import org.gdsc.globook.domain.User;
+import org.gdsc.globook.domain.type.ELanguage;
+import org.gdsc.globook.domain.type.ELoginProvider;
+import org.gdsc.globook.domain.type.EUserRole;
+import org.gdsc.globook.domain.entity.User;
 import org.gdsc.globook.presentation.request.SignInRequestDto;
 import org.gdsc.globook.presentation.request.SignUpRequestDto;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
