+++ globook_server/src/main/java/org/gdsc/globook/core/exception/GlobalErrorCode.java
@@ -37,7 +37,8 @@ public enum GlobalErrorCode implements ErrorCode{
      */
     NOT_FOUND_USER(HttpStatus.NOT_FOUND, "존재하지 않는 사용자입니다."),
     NOT_SUPPORTED_URI_ERROR(HttpStatus.NOT_FOUND, "지원하지 않는 URL 입니다."),
-    NOT_FOUND(HttpStatus.NOT_FOUND, "잘못된 접근입니다."),
+    NOT_FOUND(HttpStatus.NOT_FOUND, "존재하지 않는 데이터입니다."),
+    NOT_FOUND_BOOK(HttpStatus.NOT_FOUND, "존재하지 않는 도서입니다."),
 
     /**
      * 405 : 지원하지 않는 HTTP Method
