+++ globook_server/src/main/java/org/gdsc/globook/core/security/info/CustomUserDetails.java
@@ -7,8 +7,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.gdsc.globook.core.constant.Constants;
-import org.gdsc.globook.domain.EUserRole;
-import org.gdsc.globook.domain.User;
+import org.gdsc.globook.domain.type.EUserRole;
+import org.gdsc.globook.domain.entity.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
