+++ globook_server/src/main/java/org/gdsc/globook/presentation/controller/TestController.java
@@ -7,9 +7,9 @@ import org.gdsc.globook.core.annotation.AuthenticationPrincipal;
 import org.gdsc.globook.core.common.BaseResponse;
 import org.gdsc.globook.core.security.info.JwtToken;
 import org.gdsc.globook.core.util.JwtUtil;
-import org.gdsc.globook.domain.ELanguage;
-import org.gdsc.globook.domain.ELoginProvider;
-import org.gdsc.globook.domain.User;
+import org.gdsc.globook.domain.type.ELanguage;
+import org.gdsc.globook.domain.type.ELoginProvider;
+import org.gdsc.globook.domain.entity.User;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
